requirements.txt
# MedSafe AI - Requirements
# Activity 4.3: Deployment Preparation and Final Validation

# Core UI
streamlit>=1.32.0

# Google Generative AI / Gemini
google-generativeai>=0.5.0

# Image Processing & OCR
Pillow>=10.0.0
pytesseract>=0.3.10
opencv-python-headless>=4.9.0.80

# Data & Utilities
numpy>=1.24.0

# Logging (built-in, listed for documentation)
# logging — stdlib

# For PDF prescription support (optional)
# PyMuPDF>=1.23.0